+++ client/src/components/Button.tsx
@@ -3,7 +3,7 @@ import styled from "styled-components";
 
 // Required types and interfaces.
 import { FC } from "react";
-import Props from "../models/Props";
+import Props from "../models/props/Label";
 
 const Button: FC<Props> = ({ label }) => {
   return <Wrapper>{label}</Wrapper>;
