+++ client/src/components/Table/Table.tsx
@@ -3,7 +3,7 @@ import styled from "styled-components";
 
 // Required types and interfaces.
 import { FC } from "react";
-import Props from "../../models/Props";
+import Props from "../../models/props/Default";
 
 const Table: FC<Props> = ({ children }) => {
   return <Wrapper>{children}</Wrapper>;
