--- client/src/models/StyledProps.ts
@@ -1,3 +1,0 @@
-export default interface StyledProps {
-  filled?: boolean;
-}
