+++ client/src/models/props/Default.ts
@@ -1,0 +1,6 @@
+// Required types.
+import { ReactNode } from "react";
+
+export default interface Props {
+  children: ReactNode;
+}
