+++ client/src/models/props/Label.ts
@@ -1,0 +1,3 @@
+export default interface Props {
+  label: string;
+}
