+++ client/src/models/props/Link.ts
@@ -1,0 +1,7 @@
+// Required types.
+import { ReactNode } from "react";
+
+export default interface Props {
+  children: ReactNode;
+  path: string;
+}
